+++ frontend/src/app/util/i18n.cljs
@@ -23,14 +23,18 @@
    {:label "Français (community)" :value "fr"}
    {:label "Deutsch (community)" :value "de"}
    ;; {:label "Italiano (community)" :value "it"}
+   {:label "Norsk - Bokmål (community)" :value "nb_no"}
+   {:label "Portuguese - Brazil (community)" :value "pt_br"}
+   {:label "Polski (community)" :value "pl"}
    {:label "Русский (community)" :value "ru"}
+   {:label "Rumanian (community)" :value "ro"}
    {:label "Türkçe (community)" :value "tr"}
-   {:label "Rumanian (communit)" :value "ro"}
-   {:label "Portuguese (Brazil, community)" :value "pt_br"}
    {:label "Ελληνική γλώσσα (community)" :value "el"}
    {:label "עִבְרִית (community)" :value "he"}
    {:label "عربي/عربى (community)" :value "ar"}
-   {:label "简体中文 (community)" :value "zh_cn"}])
+   {:label "فارسی (community)" :value "fa"}
+   {:label "简体中文 (community)" :value "zh_cn"}
+   {:label "中国传统的 (community)" :value "zh_hant"}])
 
 (defn- parse-locale
   [locale]